WebExam/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
bleach==6.1.0
Markdown==3.7
Pillow==10.4.0
Werkzeug==3.0.4
gunicorn==21.2.0
